Removed old merkle methods Added Voting Key transaction V1

+++ sdk-okhttp-client/src/main/java/io/nem/symbol/sdk/infrastructure/okhttp/HashLockRepositoryOkHttpImpl.java
@@ -65,7 +65,7 @@ public class HashLockRepositoryOkHttpImpl extends AbstractRepositoryOkHttpImpl
         MapperUtils.toMosaicId(lock.getMosaicId()),
         lock.getAmount(),
         lock.getEndHeight(),
-        LockStatus.rawValueOf(lock.getStatus().getValue()),
+        LockStatus.rawValueOf(lock.getStatus().getValue().byteValue()),
         lock.getHash());
   }
