+++ sdk-vertx-client/src/main/java/io/nem/symbol/sdk/infrastructure/vertx/HashLockRepositoryVertxImpl.java
@@ -63,7 +63,7 @@ public class HashLockRepositoryVertxImpl extends AbstractRepositoryVertxImpl
         MapperUtils.toMosaicId(lock.getMosaicId()),
         lock.getAmount(),
         lock.getEndHeight(),
-        LockStatus.rawValueOf(lock.getStatus().getValue()),
+        LockStatus.rawValueOf(lock.getStatus().getValue().byteValue()),
         lock.getHash());
   }
